// assets/gulp/src/types/IManifest.ts
/**
 * Manifest
 */
export interface IManifest {
    name: string;
    short_name: string;
    description: string;
    version: string;
    author: string;
    icons: {src: string; sizes: string; type: string}[];
    theme_color: string;
    background_color: string;
    display: string;
    start_url: string;
}

// Generated by UniteJS