// assets/gulp/src/types/IUniteImageCliCommand.ts
/**
 * Unite Image Cli Command
 */
export interface IUniteImageCliCommand {
    [key: string]: any;
    command: string;
    sourceFolder?: string;
    sourceFile?: string;
    sourceFiles?: string;
    destFile: string;
    width?: number;
    height?: number;
    marginX?: number;
    marginY?: number;
    background?: string;
}

// Generated by UniteJS